requirements.txt
streamlit
pandas
streamlit-option-menu
bcrypt
pandas
langchain
langchain-community
sentence-transformers
faiss-cpu
